requirements.txt
# core
numpy==1.26.4
pillow==10.4.0
opencv-python==4.10.0.84
matplotlib==3.8.4
pandas==2.2.2
tqdm==4.66.5

# nlp utils
nltk==3.9.1

# http + yaml
requests==2.32.3
pyyaml==6.0.2

# optional: openai / anthropic clients (kept optional by env flags)
openai==1.51.2
anthropic==0.40.0
